requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
dj-database-url==3.0.0
Django==5.2.1
et_xmlfile==2.0.0
fonttools==4.58.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
xlrd==2.0.1
